+++ backend/src/main/java/com/techeerlog/global/mapper/MemberMapper.java
@@ -1,7 +1,7 @@
 package com.techeerlog.global.mapper;
 
 import com.techeerlog.member.domain.Member;
-import com.techeerlog.member.dto.MemberResponse;
+import com.techeerlog.member.dto.response.MemberResponse;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
