--- backend/src/main/java/com/techeerlog/member/dto/UniqueResponse.java
@@ -1,16 +1,0 @@
-package com.techeerlog.member.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UniqueResponse {
-
-    private boolean unique;
-
-    public UniqueResponse() {
-    }
-
-    public UniqueResponse(boolean unique) {
-        this.unique = unique;
-    }
-}
