+++ backend/src/main/java/com/techeerlog/member/dto/request/EditMemberRequest.java
@@ -1,4 +1,4 @@
-package com.techeerlog.member.dto;
+package com.techeerlog.member.dto.request;
 
 import lombok.Getter;
 import lombok.Setter;
