+++ backend/src/main/java/com/techeerlog/project/dto/ProjectMemberResponse.java
@@ -1,6 +1,6 @@
 package com.techeerlog.project.dto;
 
-import com.techeerlog.member.dto.MemberResponse;
+import com.techeerlog.member.dto.response.MemberResponse;
 import com.techeerlog.project.enums.ProjectMemberTypeEnum;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
